+++ app/javascript/entrypoints/application.js
@@ -1,0 +1,9 @@
+// This file is automatically compiled by Vite, along with any other files
+// present in this directory. You're encouraged to place your actual application logic in
+// a relevant structure within app/javascript and only use these pack files to reference
+// that code so it'll be compiled.
+console.log('Vite ⚡️ Ruby')
+import '~/stylesheets/application.scss'
+import '~/globals.js'
+import '~/setup.js'
+import '~/bootstrap.js'