+++ app/javascript/globals.js
@@ -1,0 +1,5 @@
+import jQuery from 'jquery'
+import Rails from '@rails/ujs'
+
+Object.assign(window, { $: jQuery, jQuery, Rails })
+Rails.start()